backend/requirements.txt
Django==4.2.4
django-environ==0.10.0
djangorestframework==3.14.0
PyJWT==2.8.0
requests==2.31.0
cryptography==3.4.8